feat(search): consolidate zone and hold search into one panel
Merges ClimbHoldSearchForm and ClimbZoneSearchForm into a single
ClimbSearchForm rendered under one "Holds & Zone" accordion section.
Holds and the zone rectangle now sit on the same board image so the
spatial relationship is visible at a glance.

When the user shrinks or moves the zone such that a previously-selected
filter hold falls outside, that hold is unselected on drop. The backend
zone filter requires every hold of a climb to fit inside the box, so
out-of-zone filter holds guarantee zero results — silent prune avoids
the dead-end empty result set. Both URL params update atomically in a
single updateFilters call.

Holds inside the zone stay tappable: the zone rect is pointer-events:
none, with a small explicit center handle replacing the rect-body drag.
Corner handles still resize.

Adds isHoldInsideZone helper to climb-zone-math.ts. Unit tests cover
inside / outside / on-edge / non-zero-edge-offset cases. Form-level
tests cover prune-on-zone-draw and no-prune-on-zone-clear.
