+++ packages/web/app/components/search-drawer/__tests__/accordion-search-form-handlers.test.tsx
@@ -39,16 +39,16 @@ vi.mock('../setter-name-select', () => ({
   default: () => null,
 }));
 
-vi.mock('../climb-hold-search-form', () => ({
+vi.mock('../climb-search-form', () => ({
   default: () => null,
 }));
 
 vi.mock('../search-summary-utils', () => ({
-  getQualityPanelSummary: () => '',
-  getStatusPanelSummary: () => '',
-  getUserPanelSummary: () => '',
-  getHoldsPanelSummary: () => '',
-  getZonePanelSummary: () => '',
+  getQualityPanelSummary: () => [],
+  getStatusPanelSummary: () => [],
+  getUserPanelSummary: () => [],
+  getHoldsPanelSummary: () => [],
+  getZonePanelSummary: () => [],
 }));
 
 import AccordionSearchForm from '../accordion-search-form';
