+++ packages/web/app/components/search-drawer/__tests__/climb-search-form.test.tsx
@@ -1,0 +1,153 @@
+import React from 'react';
+import { render, fireEvent, screen } from '@testing-library/react';
+import { beforeEach, describe, expect, it, vi } from 'vite-plus/test';
+import type { BoardDetails, HoldsFilter, SearchRequestPagination, ZoneBox } from '@/app/lib/types';
+import { DEFAULT_SEARCH_PARAMS } from '@/app/lib/url-utils';
+import { tFromCatalog } from '@/app/__test-helpers__/i18n-mock';
+import { gridToSvg, type BoardDimensions } from '../climb-zone-math';
+
+vi.mock('react-i18next', () => ({
+  useTranslation: (ns?: string) => ({
+    t: (key: string, options?: Record<string, unknown>) => tFromCatalog(ns, key, options),
+    i18n: { language: 'en-US' },
+  }),
+  Trans: ({ children }: { children?: React.ReactNode }) => children ?? null,
+}));
+
+const mockUpdateFilters = vi.fn();
+let mockUISearchParams: SearchRequestPagination = { ...DEFAULT_SEARCH_PARAMS };
+
+vi.mock('@/app/components/queue-control/ui-searchparams-provider', () => ({
+  useUISearchParams: () => ({
+    uiSearchParams: mockUISearchParams,
+    updateFilters: mockUpdateFilters,
+  }),
+}));
+
+vi.mock('next/navigation', () => ({
+  usePathname: () => '/kilter/1/1/1/40/list',
+}));
+
+vi.mock('@vercel/analytics', () => ({
+  track: vi.fn(),
+}));
+
+vi.mock('../../board-renderer/board-renderer', () => ({
+  default: () => null,
+}));
+
+vi.mock('../../create-climb/hold-type-picker', () => ({
+  default: () => null,
+}));
+
+vi.mock('../../create-climb/create-climb-heatmap-overlay', () => ({
+  default: () => null,
+}));
+
+vi.mock('../search-hold-filter-overlay', () => ({
+  default: () => null,
+}));
+
+vi.mock('../use-search-hold-picker', () => ({
+  useSearchHoldPicker: () => ({
+    anchorEl: null,
+    activeHoldId: null,
+    currentEntry: {},
+    handleHoldClick: vi.fn(),
+    handleFilterChange: vi.fn(),
+    handleClearAll: vi.fn(),
+    handleClose: vi.fn(),
+  }),
+}));
+
+import ClimbSearchForm from '../climb-search-form';
+
+// 144 x 156 grid mapped to 1080 x 1170 px — same dims used elsewhere in the
+// search-drawer tests so the math results are predictable.
+const dims: BoardDimensions = {
+  boardWidth: 1080,
+  boardHeight: 1170,
+  edgeLeft: 0,
+  edgeRight: 144,
+  edgeBottom: 0,
+  edgeTop: 156,
+};
+
+const holdAtGrid = (id: number, gridX: number, gridY: number) => {
+  const svgPoint = gridToSvg(gridX, gridY, dims);
+  return { id, mirroredHoldId: null, cx: svgPoint.x, cy: svgPoint.y, r: 30 };
+};
+
+// Three filter-holds: one inside the default 60% zone, two outside it (one
+// near the bottom-left corner, one near the top-right corner). The default
+// zone for these dims is { 29, 115, 31, 125 } so a hold at (60, 80) is well
+// inside, and (10, 10) / (140, 150) are well outside.
+const insideHold = holdAtGrid(101, 60, 80);
+const outsideBottomLeftHold = holdAtGrid(102, 10, 10);
+const outsideTopRightHold = holdAtGrid(103, 140, 150);
+
+const boardDetails = {
+  board_name: 'kilter',
+  layout_id: 1,
+  size_id: 1,
+  set_ids: [],
+  size_name: '12 x 12',
+  layout_name: 'Test',
+  boardWidth: dims.boardWidth,
+  boardHeight: dims.boardHeight,
+  edge_left: dims.edgeLeft,
+  edge_right: dims.edgeRight,
+  edge_bottom: dims.edgeBottom,
+  edge_top: dims.edgeTop,
+  images_to_holds: {},
+  holdsData: [insideHold, outsideBottomLeftHold, outsideTopRightHold],
+} as unknown as BoardDetails;
+
+const filterAllThreeHolds: HoldsFilter = {
+  101: { STARTING: 'include' },
+  102: { ANY: 'include' },
+  103: { FOOT: 'exclude' },
+};
+
+describe('ClimbSearchForm — zone changes prune out-of-zone holds', () => {
+  beforeEach(() => {
+    mockUpdateFilters.mockClear();
+    mockUISearchParams = { ...DEFAULT_SEARCH_PARAMS };
+  });
+
+  it('clicking Draw zone keeps only the holds inside the default zone', () => {
+    mockUISearchParams = { ...DEFAULT_SEARCH_PARAMS, holdsFilter: filterAllThreeHolds };
+    render(<ClimbSearchForm boardDetails={boardDetails} />);
+
+    const drawButton = screen.getByRole('button', { name: 'Draw zone' });
+    fireEvent.click(drawButton);
+
+    const lastCall = mockUpdateFilters.mock.calls.at(-1)?.[0] as
+      | { zoneBox: ZoneBox; holdsFilter: HoldsFilter }
+      | undefined;
+    expect(lastCall).toBeDefined();
+    expect(lastCall?.zoneBox).toEqual({ edgeLeft: 29, edgeRight: 115, edgeBottom: 31, edgeTop: 125 });
+    // Hold 101 (inside the default zone) is preserved; the two outer holds
+    // are dropped, since the backend zone filter would never return a climb
+    // that uses them anyway.
+    expect(lastCall?.holdsFilter).toEqual({ 101: { STARTING: 'include' } });
+  });
+
+  it('clicking Clear zone leaves the holdsFilter untouched', () => {
+    const existingZone: ZoneBox = { edgeLeft: 29, edgeRight: 115, edgeBottom: 31, edgeTop: 125 };
+    mockUISearchParams = {
+      ...DEFAULT_SEARCH_PARAMS,
+      holdsFilter: { 101: { STARTING: 'include' } },
+      zoneBox: existingZone,
+    };
+    render(<ClimbSearchForm boardDetails={boardDetails} />);
+
+    const clearButton = screen.getByRole('button', { name: 'Clear zone' });
+    fireEvent.click(clearButton);
+
+    const lastCall = mockUpdateFilters.mock.calls.at(-1)?.[0];
+    // Only zoneBox is cleared; holdsFilter not touched (no zone constraint
+    // means hold filters apply on their own).
+    expect(lastCall).toEqual({ zoneBox: null });
+  });
+});
