+++ packages/web/app/components/search-drawer/search-form.module.css
@@ -146,6 +146,20 @@
   touch-action: none;
 }
 
+/* Overlay SVG layered on top of BoardRenderer in the consolidated form.
+   The svg itself has pointer-events: none so taps fall through to the
+   hold click targets in BoardRenderer underneath; the handles inside opt
+   back in via pointer-events: auto on each circle. */
+.zoneOverlaySvg {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+  touch-action: none;
+}
+
 /* Search Results Footer Styles */
 .searchFooter {
   padding: 12px 16px;
